Express: added JWT to authenticate privileged operations. Angular: created login form, login button, login routes, logic state behavior, i.e data and functionality hidden when logged out. Addressed the liberal use of changedetectionref by switching items and categories to signals to automatically trigger zoneless change detection

+++ app_admin/src/app/item-listing/item-listing.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
+import { Component, OnInit, ChangeDetectorRef, signal } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Items } from '../services/items';
 import { Router, RouterLink } from '@angular/router';
+import { Authentication } from '../services/authentication';
 
 @Component({
   selector: 'app-item-listing',
@@ -12,8 +13,9 @@ import { Router, RouterLink } from '@angular/router';
 })
 export class ItemListing implements OnInit {
 
-  items: any[] = [];
-  categories: string[] = [];
+
+  items = signal<any[]>([]);
+  categories = signal<string[]>([]);
 
   // filters
   search = '';
@@ -26,39 +28,58 @@ export class ItemListing implements OnInit {
 
   constructor(private itemsService: Items, 
     private cd: ChangeDetectorRef, 
-    private router: Router) {}
+    private router: Router,
+    public auth: Authentication) {}
+    
+    isLoggedIn = false;
+
 
   ngOnInit(): void {
-    this.loadItems();
-    this.loadCategories();
+    if (this.auth.isLoggedIn()) {
+        this.loadItems();
+        this.loadCategories();
+    }
+  }
+
+  onLogout(): void {
+    this.auth.logout();
+    this.resetState();
+  }
+
+  private resetState() {
+    this.items.set([]);
+    this.categories.set([]);
   }
 
   loadItems() {
+
+    if(!this.auth.isLoggedIn()) return;
+
     const params = {
       search: this.search,
       category: this.category,
       sort: this.sort,
       order: this.order
     };
-
     
     this.itemsService.getItems(params).subscribe(res => {
-      this.items = res.items;
+      this.items.set(res.items);
       console.log('API RESPONSE:', res);
-      this.cd.detectChanges();
     });
   }
 
   loadCategories() {
-  this.itemsService.getCategories().subscribe({
-    next: (cats) => {
-      this.categories = cats;
-      this.cd.detectChanges();
-    },
-    error: (err) => {
-      console.error('Failed to load categories', err);
-    }
-  });
+
+    if(!this.auth.isLoggedIn()) return;
+
+    this.itemsService.getCategories().subscribe({
+      next: (cats) => {
+        this.categories.set(cats);
+      },
+      error: (err) => {
+        console.error('Failed to load categories', err);
+      }
+    });
   }
 
   // ======================
@@ -66,6 +87,7 @@ export class ItemListing implements OnInit {
   // ======================
 
   toggleSelection(item: any) {
+
     const index = this.selectedItems.indexOf(item);
 
     if (index >= 0) {
@@ -80,7 +102,7 @@ export class ItemListing implements OnInit {
   }
 
   toggleAll(event: any) {
-    this.selectedItems = event.target.checked ? [...this.items] : [];
+    this.selectedItems = event.target.checked ? [...this.items()] : [];
   }
 
   isAllSelected() {
@@ -93,20 +115,23 @@ export class ItemListing implements OnInit {
   // ======================
 
   deleteSelected() {
-  const ids = this.selectedItems.map(item => item._id);
-  console.log('IDs being sent:', ids);
-  this.itemsService.deleteItems(ids).subscribe({
-    next: () => {
-      this.selectedItems = [];
-      this.loadItems();
-    },
-    error: (err) => {
-      console.error('Failed to delete items', err);
-    }
-  });
-}
+    
+    if(!this.auth.isLoggedIn()) return;
 
-  editSelected() { //TODO
+    const ids = this.selectedItems.map(item => item._id);
+    console.log('IDs being sent:', ids);
+    this.itemsService.deleteItems(ids).subscribe({
+      next: () => {
+        this.selectedItems = [];
+        this.loadItems();
+      },
+      error: (err) => {
+        console.error('Failed to delete items', err);
+      }
+    });
+  }
+
+  editSelected() { // FIXME remove, unfeasible
     const item = this.selectedItems[0];
     this.router.navigate(['/items', item._id, 'edit']);
   }
@@ -116,12 +141,15 @@ export class ItemListing implements OnInit {
   }
 
   setSort(field: string) {
-  if (this.sort === field) {
-    this.order = this.order === 'asc' ? 'desc' : 'asc';
-  } else {
-    this.sort = field;
-    this.order = 'asc';
+
+    if(!this.auth.isLoggedIn()) return;
+
+    if (this.sort === field) {
+      this.order = this.order === 'asc' ? 'desc' : 'asc';
+    } else {
+      this.sort = field;
+      this.order = 'asc';
+    }
+    this.loadItems();
   }
-  this.loadItems();
-}
 }
